requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
google-adk==0.4.0
langchain==0.3.25
sentence-transformers==2.2.2
supabase==2.15.1
huggingface-hub==0.15.1
transformers==4.30.2
torch>=1.6.0
numpy<2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pdfplumber
python-docx
scikit-learn
scipy
requests
supabase-py
fastapi
uvicorn